feat: Change the place where sedding is being applied
